check server timeout behaviour
in [here](https://blog.cloudflare.com/the-complete-guide-to-golang-net-http-timeouts/), it states:

> WriteTimeout normally covers the time from the end of the request header read to the end of the response write (a.k.a. the lifetime of the ServeHTTP), by calling SetWriteDeadline at the end of readRequest.

this is actually the behaviour, and the response errors on the client
side when WriteTimeout is reached. That said, we still wait for the amount of time that it takes to process the request, regardless of the write timeout value. However, when it comes to writing the response, the server closes the conn w/o actually attempting to write it.