propogate a request level context with with writetimeout signal
since WriteTimeout normally covers the time from the end of the request header read to the end of the response write (a.k.a. the lifetime of the ServeHTTP) ([source](https://blog.cloudflare.com/the-complete-guide-to-golang-net-http-timeouts)), it makes sense to at least see the writetimeout as the request timeout.

However, request context is not signaled for cancellation when the writetimeout is reached, which means that we would still be waiting for potentially long running other operations within the request.

This PR sets the writetimeout at the request context which can be used to terminate the request when this is reached.

With the current setup, we don't yiled result much sooner:

```
➜  server git:(master) ✗ go run main.go
2020/11/23 00:41:05 [[::1]:59982,[::1]:4400]: new
2020/11/23 00:41:05 [[::1]:59982,[::1]:4400]: active
2020/11/23 00:41:08 stopped processing
2020/11/23 00:41:08 [[::1]:59982,[::1]:4400]: closed
```

```
➜  client git:(master) ✗ go run main.go
conn established [::1]:59982 [::1]:4400
conn closed [::1]:59982 [::1]:4400
2020/11/23 00:41:08 HTTP Get err: Get "http://localhost:4400": EOF
```

note that it actually makes much more sense to set request handler timeout a bit lower than the writetimeout, so that we give chance for response write. With this setup, we will have no time to write a response in case request handler hits the timeout.